Fixes tests for hyperpipes and Multilayer Perceptron
Problem with Hyperpipes was that in case of a tie, the last value in the sequence should win.
See http://www.csee.wvu.edu/~timm/tmp/r7.pdf for a description. This involved creating a new
Votes class (and accompanying tests) to tally votes and return the winner.

The multilayer perceptron had a bug where the value of the index accumulator in
data_to_input was being reassigned rather than incremented with the length of the
current domain.